Update tivo_copy_rearranged.sh
This is NOT YET TESTED. Need to upload here to download it with wget on a machine booted with jmfs-rev105-krbaker-tivotools.iso

On a drive that has no free space at the end of drive, the original version can create a final partition that is longer than the size of the destination drive.

My source drive expanded by JMFS 1.04 had this problem.

Have added a check so that the last partition is now created no longer than the size of the dest drive.

Assumes: new drive same size as old drive.

If the new drive is SMALLER than the original drive, even this won't help:)

This is a hack. Any data that (may) be in the last un-copied part of the original disk may be important. You may lose some recordings, though as it should be only a few bytes, it is unlikely.

This is NOT YET TESTED.