Now optimized partitions will not exceed size of dest drive
This is necessary for copying drives that had been previously expanded (eg by JFMS) and a drive was created with no free space at the end, and you are now copying to a drive of the same size. This is because the optimization of the partition sizes will likely (always?) create a new partition table longer than the original one. This change will effect nothing to how the script worded if you are copying to a drive larger than the original drive, and do nothing to help if the destination drive is smaller than the original (obviously).